+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Todos los cambios notables(?) del proyecto serán documentados en este archivo.
 ## [SIN PUBLICAR]
 
 ### 2019-03-26
-- agregado /CRIMSG /ACC /RACC /BMSG /SEGUIR
+- agregado /CRIMSG /ACC /RACC /BMSG /SEGUIR /RESETINV /RECOMPENSA /ENLISTAR /RETIRAR (faction) /AI1-/AI5 /AC1-/AC5
 
 ### 2019-03-25
 - agregados /RELOADNPCS /RELOADSINI /RELOADHECHIZOS /RELOADOBJ /ROL /AEMAIL /ANAME /APASS /BLOQ /TRIGGER